# feat(json): reduce string int decoding allocations