Benchstat comparison:
```
name                 old time/op    new time/op    delta
DecodeStringInt32-4    96.3ns ± 2%    55.0ns ± 1%   -42.87%  (p=0.000 n=14+15)
DecodeStringInt64-4    97.1ns ± 2%    52.5ns ± 1%   -45.89%  (p=0.000 n=14+15)

name                 old alloc/op   new alloc/op   delta
DecodeStringInt32-4     16.0B ± 0%      0.0B       -100.00%  (p=0.000 n=15+15)
DecodeStringInt64-4     16.0B ± 0%      0.0B       -100.00%  (p=0.000 n=15+15)

name                 old allocs/op  new allocs/op  delta
DecodeStringInt32-4      1.00 ± 0%      0.00       -100.00%  (p=0.000 n=15+15)
DecodeStringInt64-4      1.00 ± 0%      0.00       -100.00%  (p=0.000 n=15+15)
```